fix(session): report actual jj journal command
Share jj command formatting between execution failures and empty-output JournalError::CommandFailed paths so diagnostics include the injected --no-pager and --color=never arguments. Move jj journal unit tests into a sibling test module to satisfy the monolith gate, and bump the workspace patch version required by pre-commit.\n\nVerification:\n- cargo test -p csa-session jj_journal_tests::empty_operation_id_reports_actual_jj_command\n- cargo test -p csa-process tool_liveness::tests::daemon_pid_is_alive_accepts_legacy_pid_with_session_id_context\n- cargo test --workspace\n- cargo clippy --workspace
